# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/alimento/": {
      "get": {
        "tags": [
          "alimento-controller"
        ],
        "operationId": "getAll",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alimento"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "alimento-controller"
        ],
        "operationId": "addAlimento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Alimento"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response"
          }
        }
      }
    },
    "/alimento/setAccesible": {
      "post": {
        "tags": [
          "alimento-controller"
        ],
        "operationId": "setAccesible",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccesibleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response"
          }
        }
      }
    },
    "/alimento/cargarTablas": {
      "post": {
        "tags": [
          "alimento-controller"
        ],
        "operationId": "cargarTablas",
        "responses": {
          "200": {
            "description": "default response"
          }
        }
      }
    },
    "/alimento/{alimentoId}": {
      "get": {
        "tags": [
          "alimento-controller"
        ],
        "operationId": "getFrutaById",
        "parameters": [
          {
            "name": "alimentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Alimento"
                }
              }
            }
          }
        }
      }
    },
    "/receta/": {
      "get": {
        "tags": [
          "receta-controller"
        ],
        "operationId": "getAll_1",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Receta"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "receta-controller"
        ],
        "operationId": "addReceta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Receta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response"
          }
        }
      }
    },
    "/receta/{idReceta}": {
      "delete": {
        "tags": [
          "receta-controller"
        ],
        "operationId": "deleteReceta",
        "parameters": [
          {
            "name": "idReceta",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response"
          }
        }
      }
    },
    "/receta/modificar": {
      "post": {
        "tags": [
          "receta-controller"
        ],
        "operationId": "modificarReceta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Receta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response"
          }
        }
      }
    },
    "/unidades/": {
      "get": {
        "tags": [
          "unidades-controller"
        ],
        "operationId": "getAll_2",
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnidadDeMedida"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "unidades-controller"
        ],
        "operationId": "addUnidad",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnidadDeMedida"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Alimento": {
        "type": "object",
        "properties": {
          "numero": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "clasificacion": {
            "type": "string"
          },
          "genero_especie_variedad": {
            "type": "string"
          },
          "energia_kJ": {
            "type": "number",
            "format": "double"
          },
          "agua": {
            "type": "number",
            "format": "double"
          },
          "proteina": {
            "type": "number",
            "format": "double"
          },
          "grasa_total": {
            "type": "number",
            "format": "double"
          },
          "carbohidrato_total": {
            "type": "number",
            "format": "double"
          },
          "carbohidrato_disponible": {
            "type": "number",
            "format": "double"
          },
          "fibra_dietetica": {
            "type": "number",
            "format": "double"
          },
          "ceniza": {
            "type": "number",
            "format": "double"
          },
          "sodio": {
            "type": "number",
            "format": "double"
          },
          "potasio": {
            "type": "number",
            "format": "double"
          },
          "calcio": {
            "type": "number",
            "format": "double"
          },
          "fosforo": {
            "type": "number",
            "format": "double"
          },
          "hierro": {
            "type": "number",
            "format": "double"
          },
          "zinc": {
            "type": "number",
            "format": "double"
          },
          "tiamina": {
            "type": "number",
            "format": "double"
          },
          "rivoflavina": {
            "type": "number",
            "format": "double"
          },
          "niacina": {
            "type": "number",
            "format": "double"
          },
          "vitamina_c": {
            "type": "number",
            "format": "double"
          },
          "acidos_grasos_saturados": {
            "type": "number",
            "format": "double"
          },
          "acidos_grasos_monoinsaturados": {
            "type": "number",
            "format": "double"
          },
          "acidos_grasos_poliinsaturados": {
            "type": "number",
            "format": "double"
          },
          "colesterol": {
            "type": "number",
            "format": "double"
          },
          "esAccesible": {
            "type": "boolean"
          }
        }
      },
      "AccesibleDTO": {
        "type": "object",
        "properties": {
          "numero": {
            "type": "integer",
            "format": "int32"
          },
          "esAccesible": {
            "type": "boolean"
          }
        }
      },
      "Ingrediente": {
        "type": "object",
        "properties": {
          "idIngrediente": {
            "type": "integer",
            "format": "int32"
          },
          "alimento": {
            "$ref": "#/components/schemas/Alimento"
          },
          "cantidad": {
            "type": "number",
            "format": "double"
          },
          "unidadDeMedida": {
            "$ref": "#/components/schemas/UnidadDeMedida"
          }
        }
      },
      "Receta": {
        "type": "object",
        "properties": {
          "idReceta": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "ingredientes": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingrediente"
            }
          },
          "instrucciones": {
            "type": "string"
          }
        }
      },
      "UnidadDeMedida": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          }
        }
      }
    }
  }
}