# src/main/java/com/mau/spring/model/Receta.java
package com.mau.spring.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.persistence.*;
import java.util.Set;

@AllArgsConstructor
@NoArgsConstructor
@Data
@Entity
public class Receta
{
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer idReceta;
    private String nombre;
    private String descripcion;
    @OneToMany(cascade = {CascadeType.ALL})
    private Set<Ingrediente> ingredientes;

    private Integer cantidadPlatos;
    private String instrucciones;

    private double energia_kJ;
    private double agua;
    private double proteina;
    private double grasa_total;
    private double carbohidrato_total;
    private double carbohidrato_disponible;
    private double fibra_dietetica;
    private double ceniza;
    private double sodio;
    private double potasio;
    private double calcio;
    private double fosforo;
    private double hierro;
    private double zinc;
    private double tiamina;
    private double rivoflavina;
    private double niacina;
    private double vitamina_c;
    private double acidos_grasos_saturados;
    private double acidos_grasos_monoinsaturados;
    private double acidos_grasos_poliinsaturados;
    private double colesterol;
}